+++ src/database/queries/authorization/hasSourcePermission.sql
@@ -1,7 +1,11 @@
-SELECT has_source_permission(
-	:userKeycloakUserId,
-	:isAdministrator,
-	:sourceId,
-	:permission::permission_grant_verb_t,
-	:scope::permission_grant_entity_type_t
+SELECT exists(
+	SELECT 1
+	FROM
+		permitted_source_ids(
+			:userKeycloakUserId,
+			:isAdministrator,
+			:permission::permission_grant_verb_t,
+			:scope::permission_grant_entity_type_t
+		) AS permitted_sources
+	WHERE permitted_sources.id = :sourceId
 ) AS "hasPermission";