+++ src/database/queries/authorization/hasChangemakerFieldValuePermission.sql
@@ -1,7 +1,11 @@
-SELECT has_changemaker_field_value_permission(
-	:userKeycloakUserId,
-	:isAdministrator,
-	:changemakerFieldValueId,
-	:permission::permission_grant_verb_t,
-	:scope::permission_grant_entity_type_t
+SELECT exists(
+	SELECT 1
+	FROM
+		permitted_changemaker_field_value_ids(
+			:userKeycloakUserId,
+			:isAdministrator,
+			:permission::permission_grant_verb_t,
+			:scope::permission_grant_entity_type_t
+		) AS permitted_field_values
+	WHERE permitted_field_values.id = :changemakerFieldValueId
 ) AS "hasPermission";