+++ src/database/queries/permissionGrants/selectWithPagination.sql
@@ -2,11 +2,12 @@ WITH
 	candidate_entries AS MATERIALIZED (
 		SELECT permission_grants.*
 		FROM permission_grants
-		WHERE can_manage_permission_grant(
-			:authContextKeycloakUserId,
-			:authContextIsAdministrator,
-			permission_grants.*::permission_grants
-		)
+			INNER JOIN
+				permitted_permission_grant_ids(
+					:authContextKeycloakUserId,
+					:authContextIsAdministrator
+				) AS permitted_grants
+				ON permission_grants.id = permitted_grants.id
 	),
 
 	entry_count AS (