+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.39.0 2026-06-24
+
 ### Changed
 
 - Listing opportunities and bulk upload tasks no longer slows down as the number of permission grants grows.